size_t is unsigned; changing index from int to size_t is not always a good idea

+++ dev/lib/geomatics/DiscCorr.cpp
@@ -2725,7 +2725,7 @@ try {
 
       // GF only slip - compute WL stats on both sides
       Stats<double> futureStats,pastStats;
-      size_t k = i;
+      int k = i;
       // fill future
       while(k < size() && futureStats.N() < N) {
          if(spdvector[k].flag & OK)                  // data is good
