+++ dev/src/RinexObsID.cpp
@@ -127,7 +127,7 @@ namespace gpstk
    // described in the Rinex 3 specification.
    bool isValidRinexObsID(const std::string& strID)
    {
-      size_t i(strID.length()-3);
+      int i(strID.length()-3);
       if(i < 0 || i > 1)
          return false;
